requirements.txt
click
imutils
opencv-python
rembg
waitress
paddlepaddle 
paddleocr 
numpy 
pillow 
matplotlib 
pandas 
fastapi 
uvicorn 
scipy
